+++ packages/core/permission/index.ts
@@ -6,7 +6,7 @@ export {
     XmIfSessionDirective,
 } from './src/directives/xm-if-session.directive';
 
-export { PermissionContext, PermissionDirective } from './src/permission.directive';
+export { PermissionContext, PermissionDirective } from './src/directives/permission.directive';
 export { PermissionGuard, PermissionGuardData } from './src/permission.guard';
 export { XmPermissionService, PermissionCheckStrategy, SUPER_ADMIN } from './src/xm-permission.service';
 export { XmPermissionModule } from './src/xm-permission.module';
