+++ packages/core/permission/src/xm-permission.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { PermissionDirective } from './permission.directive';
+import { PermissionDirective } from './directives/permission.directive';
 import { XmIfSessionDirective } from './directives/xm-if-session.directive';
 
 @NgModule({
